+++ eslint.config.mjs
@@ -1,48 +1,37 @@
-import typescriptEslint from "@typescript-eslint/eslint-plugin";
-import tsParser from "@typescript-eslint/parser";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-    baseDirectory: __dirname,
-    recommendedConfig: js.configs.recommended,
-    allConfig: js.configs.all
-});
+import tsParser from '@typescript-eslint/parser';
+import prettierConfig from 'eslint-config-prettier';
 
-export default [{
-    ignores: ["**/node_modules"],
-}, ...compat.extends(
-    "eslint:recommended",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
-    "prettier",
-), {
-    plugins: {
-        "@typescript-eslint": typescriptEslint,
-    },
-
+export default tseslint.config(
+  eslint.configs.recommended,
+  tseslint.configs.recommended,
+  prettierConfig,
+  {
+    ignores: ['**/node_modules', '**/.sf', '**/.sfdx'],
     languageOptions: {
-        parser: tsParser,
-        ecmaVersion: "latest",
-        sourceType: "module",
+      parser: tsParser,
+      ecmaVersion: 'latest',
+      sourceType: 'module',
     },
 
     rules: {
-        "no-console": "warn",
-        "@typescript-eslint/naming-convention": "warn",
-        semi: "warn",
+      'no-console': 'warn',
+      '@typescript-eslint/naming-convention': 'warn',
+      semi: 'warn',
 
-        "@typescript-eslint/no-unused-vars": ["error", {
-            argsIgnorePattern: "^_",
-            varsIgnorePattern: "^_",
-        }],
+      '@typescript-eslint/no-unused-vars': [
+        'error',
+        {
+          argsIgnorePattern: '^_',
+          varsIgnorePattern: '^_',
+        },
+      ],
 
-        "@typescript-eslint/no-explicit-any": "warn",
-        curly: "warn",
-        eqeqeq: "warn",
+      '@typescript-eslint/no-explicit-any': 'warn',
+      curly: 'warn',
+      eqeqeq: 'warn',
     },
-}];
+  },
+);