+++ rolldown.config.ts
@@ -1,0 +1,87 @@
+import { defineConfig } from 'rolldown';
+
+// rolldown plugins
+import nodePolyfills from '@rolldown/plugin-node-polyfills';
+
+// rollup plugins
+import copy from 'rollup-plugin-copy';
+import postcss from 'rollup-plugin-postcss';
+import { defineRollupSwcOption, swc } from 'rollup-plugin-swc3';
+
+const getSwcOptions = (dirPath: string) =>
+  defineRollupSwcOption({
+    include: /\.[mc]?[jt]sx?$/,
+    exclude: 'node_modules',
+    tsconfig: production ? `${dirPath}/tsconfig.json` : `${dirPath}/tsconfig-dev.json`,
+    jsc: {
+      transform: { useDefineForClassFields: false },
+      minify: {
+        compress: production,
+        mangle: production
+          ? {
+              keepClassNames: true,
+            }
+          : false,
+      },
+    },
+  });
+
+const production = process.env.NODE_ENV === 'production';
+console.log('Package mode:', production ? 'production' : 'development');
+export default defineConfig([
+  {
+    input: './lana/src/Main.ts',
+    output: {
+      format: 'esm',
+      dir: './lana/out',
+      chunkFileNames: 'lana-[name].js',
+      sourcemap: false,
+    },
+    platform: 'node',
+    external: ['vscode'],
+    resolve: { tsconfigFilename: production ? './lana/tsconfig.json' : './lana/tsconfig-dev.json' },
+    plugins: [swc(getSwcOptions('./lana'))],
+  },
+  {
+    input: { bundle: './log-viewer/modules/Main.ts' },
+    output: [
+      {
+        format: 'esm',
+        dir: './log-viewer/out',
+        chunkFileNames: 'log-viewer-[name].js',
+        sourcemap: false,
+      },
+    ],
+    platform: 'browser',
+    keepNames: true,
+    resolve: {
+      tsconfigFilename: production
+        ? './log-viewer/tsconfig.json'
+        : './log-viewer/tsconfig-dev.json',
+    },
+    transform: { typescript: {} },
+    plugins: [
+      nodePolyfills(),
+      postcss({
+        extensions: ['.css', '.scss'],
+        minimize: true,
+      }),
+      swc(getSwcOptions('./log-viewer')),
+      copy({
+        hook: 'closeBundle',
+        targets: [
+          {
+            src: [
+              'log-viewer/out/*',
+              'log-viewer/index.html',
+              'lana/certinia-icon-color.png',
+              'node_modules/@vscode/codicons/dist/codicon.ttf',
+            ],
+            dest: 'lana/out',
+          },
+          { src: ['CHANGELOG.md', 'LICENSE.txt', 'README.md'], dest: 'lana' },
+        ],
+      }),
+    ],
+  },
+]);